desktop-app/requirements.txt
pynput
sounddevice
soundfile
requests
numpy
ttkthemes
mcp
google-api-python-client
google-auth-oauthlib
litellm
openai-whisper
torch
torchaudio
ffmpeg
jira
python-dotenv